Validate organizations (#120)
* Adds endpoint for EIN check in controller

* Adds Exception responses based on checkEIN function fetch request response

* Creates Propublica Org type; Changes checkEin function name to getProPublicaOrg; throws NotFoundException if org not found

+++ server/src/organizations/organizations.service.ts
@@ -1,19 +1,41 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateOrganizationDto } from './dto/create-organization.dto';
 import { UpdateOrganizationDto } from './dto/update-organization.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, DeleteResult } from 'typeorm';
 import { Organization } from './entities/organization.entity';
+import fetch from 'node-fetch';
+
+export type PropublicaOrg = {
+  ein: number;
+  name: string;
+};
 
 @Injectable()
 export class OrganizationsService {
   constructor(
-    @InjectRepository(Organization)
-    private organizationsRepository: Repository<Organization>,
+    @InjectRepository(Organization) private organizationsRepository: Repository<Organization>,
   ) {}
 
   async create(createOrganizationDto: CreateOrganizationDto): Promise<Organization> {
-    return await this.organizationsRepository.save(createOrganizationDto);
+    return this.organizationsRepository.save(createOrganizationDto);
+  }
+
+  async getProPublicaOrg(ein: number): Promise<PropublicaOrg> {
+    try {
+      const res = await fetch(
+        `https://projects.propublica.org/nonprofits/api/v2/organizations/${ein}.json`,
+      );
+      const org = await res.json();
+      if (org) {
+        return {
+          ein: org.organization.ein,
+          name: org.organization.name,
+        };
+      }
+    } catch (err) {
+      throw new NotFoundException('Organization Not Found in Propublica API');
+    }
   }
 
   findAll(): Promise<Organization[]> {
@@ -25,11 +47,19 @@ export class OrganizationsService {
   }
 
   async update(id: number, updateOrganizationDto: UpdateOrganizationDto): Promise<Organization> {
-    await this.organizationsRepository.update(id, updateOrganizationDto);
-    return this.organizationsRepository.findOne(id);
+    const org = await this.findOne(id);
+    if (!org) {
+      throw new NotFoundException('Organization Not Found');
+    }
+    Object.assign(org, updateOrganizationDto);
+    return this.organizationsRepository.save(org);
   }
 
-  remove(id: number): Promise<DeleteResult> {
+  async remove(id: number): Promise<DeleteResult> {
+    const org = await this.findOne(id);
+    if (!org) {
+      throw new NotFoundException('Organization Not Found');
+    }
     return this.organizationsRepository.delete(id);
   }
 }
